requirements-api.txt
fastapi
uvicorn[standard]
python-dotenv
requests
psycopg[binary]
python-dotenv